requirements.txt
fastapi
uvicorn[standard]
openai
chromadb
python-multipart
pydantic
pydantic-settings
python-dotenv
PyPDF2
python-docx
langchain-text-splitters
httpx
streamlit
requests